refactor(test): extract addSpaceAdmin to SpaceClient, extract joinTeam and addTeamAdmin to TeamClient
